requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg[binary]==3.2.7
jinja2==3.1.6
python-multipart==0.0.20
python-dotenv==1.1.0
